Rename spritetypes to json2spriteanimation
Temporary name, we should make a file for storing
data/resources loading functions and their like.
